refactor: fetching schedule with useSwrHook
